fix(overlaybd): validate index bounds before loading

+++ storage/overlaybd/src/lsmt/file/readonly.rs
@@ -31,12 +31,9 @@ impl LSMTReadOnlyFile {
 
         let trailer = verify_ht(&file, true, file_size).await?;
 
-        let mappings = load_index_and_reset_tags(
-            &file,
-            trailer.index_offset.get(),
-            trailer.index_size.get() as usize,
-        )
-        .await?;
+        let mappings =
+            load_index_and_reset_tags(&file, trailer.index_offset.get(), trailer.index_size.get())
+                .await?;
 
         let index = Arc::new(ReadOnlyIndex::new(mappings));
         let uuid = parse_uuid_field(&trailer.uuid).unwrap_or_else(Uuid::nil);
